backend/requirements.txt
# Dependências principais
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
pypdfium2>=4.30,<5.0
pillow>=10.0,<11.0
pdfplumber>=0.11,<1.0
pymupdf>=1.23,<2.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
xlrd>=2.0,<3.0
matplotlib>=3.8,<4.0

# Dependências opcionais
# ortools - necessário apenas para estratégia CP-SAT
ortools

# reportlab - necessário apenas para geração de PDFs de escalas
reportlab

# Dependências adicionais
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
google-auth==2.34.0
python-multipart==0.0.12
requests>=2.31,<3.0
bcrypt>=4.0,<5.0

pydantic==2.10.6
sqlmodel==0.0.22
alembic==1.13.2
psycopg[binary]==3.2.3
psycopg2-binary  # Compatibilidade com SQLAlchemy 2.0.45
redis==5.0.8
arq==0.26.1
httpx==0.27.2
boto3==1.35.36
resend>=2.0.0
